iox-#10 Simplify module structure
