--- src/pb/mod.rs
@@ -1,10 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-FileCopyrightText: © Contributors to the iceoryx-rs project
-// SPDX-FileContributor: Mathias Kraus
-
-mod publisher;
-mod sample;
-
-pub use publisher::{InactivePublisher, Publisher, PublisherBuilder};
-
-pub use sample::SampleMut;
